+++ day02_test.go
@@ -6,7 +6,7 @@ import (
 )
 
 func TestExecuteDay02Intcode(t *testing.T) {
-	for codeStr, expResult := range map[string][]int{
+	for codeStr, expResult := range map[string][]int64{
 		"1,0,0,0,99":          {2, 0, 0, 0, 99},
 		"2,3,0,3,99":          {2, 3, 0, 6, 99},
 		"2,4,4,5,99,0":        {2, 4, 4, 5, 99, 9801},
